Navigation Bar #2
Second attempt at creating a navigation bar. I used color code detector to set the color of the bar. I used text align and word spacing to create a more ordered look. I could not center the links on the bar so I did it manually by changing the border radius. I used in-line elements (a picture and a quote) to place them side by side. My navigation bar is a block element.